+++ src/anomalib/engine/accelerator/__init__.py
@@ -1,0 +1,8 @@
+"""Accelerator for Lightning Trainer."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .xpu import XPUAccelerator
+
+__all__ = ["XPUAccelerator"]